Some refactoring, removal of dead and redundent code, simplified entry syncing
