+++ state/entrySyncing.go
@@ -6,60 +6,14 @@ package state
 
 import (
 	"github.com/FactomProject/factomd/common/constants"
-	"github.com/FactomProject/factomd/common/messages"
 )
-
-func (s *State) setTimersMakeRequests() {
-	now := s.GetTimestamp()
-
-	// If we have no Entry Blocks in our queue, reset our timer.
-	if len(s.MissingEntryBlocks) == 0 || s.MissingEntryBlockRepeat == nil {
-		s.MissingEntryBlockRepeat = now
-	}
-
-	// If our delay has been reached, then ask for some missing Entry blocks
-	// This is a replay, because sometimes requests are ignored or lost.
-	if now.GetTimeSeconds()-s.MissingEntryBlockRepeat.GetTimeSeconds() > 5 {
-		s.MissingEntryBlockRepeat = now
-
-		for _, eb := range s.MissingEntryBlocks {
-			eBlockRequest := messages.NewMissingData(s, eb.ebhash)
-			s.NetworkOutMsgQueue() <- eBlockRequest
-		}
-	}
-
-	if len(s.MissingEntries) == 0 {
-		s.MissingEntryRepeat = nil
-		s.EntryDBHeightComplete = s.EntryBlockDBHeightComplete
-		s.EntryDBHeightComplete = s.EntryDBHeightComplete
-	} else {
-		if s.MissingEntryRepeat == nil {
-			s.MissingEntryRepeat = now
-		}
-
-		// If our delay has been reached, then ask for some missing Entry blocks
-		// This is a replay, because sometimes requests are ignored or lost.
-		if now.GetTimeSeconds()-s.MissingEntryRepeat.GetTimeSeconds() > 5 {
-			s.MissingEntryRepeat = now
-
-			for i, eb := range s.MissingEntries {
-				if i > 200 {
-					// Only send out 200 requests at a time.
-					break
-				}
-				entryRequest := messages.NewMissingData(s, eb.entryhash)
-				s.NetworkOutMsgQueue() <- entryRequest
-			}
-		}
-	}
-}
 
 func (s *State) syncEntryBlocks() {
 	// All done is true, and as long as it says true, we walk our bookmark forward.  Once we find something
 	// missing, we stop moving the bookmark, and rely on caching to keep us from thrashing the disk as we
 	// review the directory block over again the next time.
 	alldone := true
-	for s.EntryBlockDBHeightProcessing < s.GetHighestCompletedBlk() && len(s.MissingEntryBlocks) < 10 {
+	for s.EntryBlockDBHeightProcessing < s.GetHighestCompletedBlk() && len(s.MissingEntryBlocks) < 300 {
 		dbstate := s.DBStates.Get(int(s.EntryBlockDBHeightProcessing))
 
 		if dbstate == nil {
@@ -105,7 +59,7 @@ func (s *State) syncEntryBlocks() {
 
 func (s *State) syncEntries(eights bool) {
 
-	for s.EntryDBHeightProcessing < s.GetHighestCompletedBlk() && len(s.MissingEntries) < 10 {
+	for s.EntryDBHeightProcessing < s.GetHighestCompletedBlk() && len(s.MissingEntries) < 300 {
 		dbstate := s.DBStates.Get(int(s.EntryDBHeightProcessing))
 
 		if dbstate == nil {
@@ -179,8 +133,6 @@ func (s *State) syncEntries(eights bool) {
 // called.
 
 func (s *State) catchupEBlocks() {
-
-	s.setTimersMakeRequests()
 
 	// If we still have blocks that we are asking for, then let's not add to the list.
