+++ state/loadDatabase.go
@@ -61,10 +61,6 @@ func LoadDatabase(s *State) {
 						time.Sleep(10 * time.Millisecond)
 					}
 				}
-				s.EntryDBHeightComplete = uint32(i)
-				s.EntryDBHeightProcessing = uint32(i)
-				s.EntryBlockDBHeightComplete = s.EntryDBHeightComplete
-				s.EntryBlockDBHeightProcessing = s.EntryDBHeightProcessing
 			} else {
 				// os.Stderr.WriteString(fmt.Sprintf("%20s Last Block in database: %d\n", s.FactomNodeName, i))
 				break
